Added a demonstration that using a grayscale image as alphachannel texture.
In this case, the grayscale image is used as a template. User can
choose arbitary color to fill into black area of the grayscale
image. White area is treated as tranparent.

Technique demonstrated here can also be applied to font rendering with
Freetype2. In Freetype2, a grayscale image is generated for each
selected character (glyph). Those grayscale images can be rendered
with method used in this code (gl_texture_grayscale.c).